Update VPQME
Includes equations from 825 to 876, it makes explicit the calculations of eigenvalues and eigenvectors, the tilt angle \theta shown in the paper PRB2011 is shown from the equation 839. Changes in the matrix representation of |1><1| and \sigma_y in the calculations of the expected values. It has effects too on the equations 914-929.