Update VPQME
This document includes equations aligned until equation 421, furthermore it solves the following parts:
- Eqs. 25-36 would ideally be in the same block, so do insert -> formatting -> new page before Eq. 25 
The equations belong to one block in the page 3.
- For Eq. 83-87, why are the equals signs not aligned? Please make this one block.

These equations now belong to one block but I had to do the equation 86 in the smallest font in order to use only one line.
- Why are Eqs. 232-234 so squished?

I used the normal font, in order to use only one line then I used negative spaces to reduce the length of the equations.

- WHY IS Eq. 309 so squished ???
- Eq. 388-390 don't need to be so squished. Keep in mind in VPQME_PAPER.pdf, I found a way to make rho^thermal take less space.

I changed \rho^Thermal to \rho_B in the document and defined that these density matrix are equivalent. I did in the text before the equation 340.
- Eqs. 408-409 can be bigger
I made them bigger.