Update VPQME
This document contains the quantities dependent of time with this dependent shown until the equation 439 of the new document. Also it includes a definition in equation 436 that explains the time dependence of the operators B and the time evolution. 

Furthermore the equations from 440 to 458 are fixed following the definition explained.