Updated VPQME_PAPER
This commit include changes in 38, 39 and 169 because they didn't match with the equation 36, also is shown the definitions 154 and 155 of the new document to proof the master equation of PRB2011. the notes between 161 and 162 can be used to track the algebra. Finally I added the equations from 162 to 166, the equation 166 has the same final form thatn PRB2011.