Update VPQME
This document contains the equations from 1401 to 1452 which shows a form to calculate the coefficients vk using the expected value of an operator written in terms of H_0 and H_I in the equation 1428 of the new document. Also includes the terms v2 and v3 present the paper of Decoster and Kimura et al. for the second quantum Bogoliubov inequality in the equations 1429 to 1439 of the new document. Also the document defines E_free,MNs in the equation 1448 to denote the strong form of the Quantum Bogoliubov inequality of order Mn and the equation 1452 of the new document that shows the weak form of the Quantum Bogoliubov inequality that was used by Kimura et al.